#304 fix(routine): Correct ID retrieval type and DAO method usage
This commit addresses data handling inconsistencies related to Routines. The `NotificationActionReceiver` has been updated to correctly retrieve the routine ID as an `Int` from the intent extras, fixing a type mismatch where it was previously accessed as a `Long`.

Additionally, the `HomeRepositoryImpl` now calls the correct DAO method `getRoutineById` instead of `getRoutine` to ensure proper data retrieval.

**Key Changes:**

*   **`NotificationActionReceiver.kt`**:
    *   Changed the retrieval of `KEY_REMINDER_ID` from `getLongExtra` to `getIntExtra`.
*   **`HomeRepositoryImpl.kt`**:
    *   Updated the `getRoutine` function to invoke `routineDao.getRoutineById` instead of `routineDao.getRoutine`.

+++ app/src/main/kotlin/com/rahim/data/reminder/notification/NotificationActionReceiver.kt
@@ -21,7 +21,7 @@ class NotificationActionReceiver : BroadcastReceiver(),KoinComponent {
   override fun onReceive(context: Context, intent: Intent) {
     if (intent.action == ACTION_CANCEL_NOTIFICATION) {
       val notificationId = intent.getIntExtra(NOTIFICATION_ID, -1)
-      val routineId = intent.getLongExtra(KEY_REMINDER_ID, -1)
+      val routineId = intent.getIntExtra(KEY_REMINDER_ID, -1)
       if (notificationId != -1) {
         with(NotificationManagerCompat.from(context)) {
           if (ActivityCompat.checkSelfPermission(
